92 move searchbar under appbar
