+++ src/Components/Appbar/Searchbar/Searchbar.test.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import { Searchbar } from '../../';
+import { Searchbar } from '../../../';
 
 import ShallowRenderer from 'react-test-renderer/shallow';
 const renderer = new ShallowRenderer();
